cli_folder/requirements.txt
camelot-py==0.10.1
PyPDF2==1.26.0
pdf2image==1.16.3
pytesseract==0.3.10
pandas==1.5.3
torch==2.0.1
transformers==4.30.2
opencv-python-headless==4.7.0.72
numpy==1.24.3
Pillow==9.5.0
ghostscript==0.7
timm==0.6.13
torchvision==0.15.2
datasets==2.13.0
matplotlib==3.7.1